Moved signal error test and added error when opening JSON files
