+++ tests/UnitTests/PhasorDynamics/CMakeLists.txt
@@ -49,16 +49,13 @@ add_executable(test_phasor_system runSystemTests.cpp)
 target_link_libraries(test_phasor_system GridKit::definitions
                                          GridKit::phasor_dynamics_components
                                          GridKit::phasor_dynamics_components_dependency_tracking)
+configure_file(${CMAKE_CURRENT_SOURCE_DIR}/ThreeBusBasicBad.json
+  ${CMAKE_CURRENT_BINARY_DIR}/ThreeBusBasicBad.json
+  COPYONLY)
 
 add_executable(test_phasor_system_single_component runSystemSingleComponentTests.cpp)
 target_link_libraries(test_phasor_system_single_component GridKit::phasor_dynamics_components
                                                           GridKit::phasor_dynamics_components_dependency_tracking)
-
-add_executable(test_phasor_system_model runSystemModelTests.cpp)
-target_link_libraries(test_phasor_system_model GridKit::phasor_dynamics_components)
-configure_file(${CMAKE_CURRENT_SOURCE_DIR}/ThreeBusBasicBad.json
-  ${CMAKE_CURRENT_BINARY_DIR}/ThreeBusBasicBad.json
-  COPYONLY)
 
 add_test(NAME PhasorDynamicsBusTest COMMAND test_phasor_bus)
 add_test(NAME PhasorDynamicsBranchTest COMMAND test_phasor_branch)
@@ -66,9 +63,9 @@ add_test(NAME PhasorDynamicsGenrouTest COMMAND test_phasor_genrou)
 add_test(NAME PhasorDynamicsGovernorTgov1Test COMMAND test_phasor_governortgov1)
 add_test(NAME PhasorDynamicsGenClassicalTest COMMAND test_phasor_gen_classical)
 add_test(NAME PhasorDynamicsLoadTest COMMAND test_phasor_load)
-add_test(NAME PhasorDynamicsSystemTest COMMAND test_phasor_system)
+add_test(NAME PhasorDynamicsSystemTest COMMAND test_phasor_system
+    WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR})
 add_test(NAME PhasorDynamicsSystemSingleComponentTest COMMAND test_phasor_system_single_component)
-add_test(NAME PhasorDynamicsSystemModelTest COMMAND test_phasor_system_model)
 
 install(TARGETS test_phasor_bus
                 test_phasor_branch
