--- tests/UnitTests/PhasorDynamics/runSystemModelTests.cpp
@@ -1,14 +1,0 @@
-#include "SystemModelTests.hpp"
-
-int main()
-{
-  using namespace GridKit;
-  using namespace GridKit::Testing;
-
-  TestingResults   result;
-  SystemModelTests test;
-
-  result += test.signalError();
-
-  return result.summary();
-}
